Feature: декомпозировать компонент AvatarUpload
