+++ src/features/dashboard/components/avatarUpload/AvatarUpload.module.scss
@@ -12,33 +12,4 @@
       opacity: 0.9;
     }
   }
-
-  &__upload {
-    &_btn {
-      color: var(--colors-black, #111);
-      text-align: center;
-      font-family: Roboto;
-      font-size: 20px;
-      font-style: normal;
-      font-weight: 400;
-      line-height: 28px;
-    }
-
-    &_wrapper {
-      display: flex;
-      flex-direction: column;
-      justify-content: end;
-      row-gap: 16px;
-      align-items: start;
-    }
-
-    &_subtitle {
-      color: var(--colors-gray-k, #6b6b6b);
-      font-family: Roboto;
-      font-size: 16px;
-      font-style: normal;
-      font-weight: 400;
-      line-height: 24px; /* 150% */
-    }
-  }
 }
